requirements.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
arrow==0.7.0
Babel==2.3.3
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
cleo==0.4.1
click==6.6
configparser==3.5.0b2
docker-py==1.7.0
docopt==0.6.2
docutils==0.12
fake-factory==0.5.7
Flask==0.10.1
humanfriendly==1.44.6
imagesize==0.7.0
inflection==0.3.1
irma==1.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
lazy-object-proxy==1.2.2
MarkupSafe==0.23
marshmallow==2.7.3
orator==0.8.1
pika==0.10.0
pony==0.6.5
Project==0.1
psutil==4.1.0
psycopg2==2.6.1
pyaml==15.8.2
Pygments==2.1.3
pylev==1.3.0
PyMySQL==0.7.2
python-dateutil==2.5.3
python-magic==0.4.11
pytz==2016.3
PyYAML==3.11
requests==2.9.1
-e git+https://github.com/deepgram/sidomo.git@b70825d9017eb0c2c6b6389345cccbcbd52cf669#egg=sidomo
simplejson==3.8.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.4
SQLAlchemy==1.0.12
transaction==1.4.4
websocket-client==0.37.0
Werkzeug==0.11.5
wrapt==1.10.8
